tag:运营记录

+++ galaxy-sop-api/src/main/java/com/galaxyinternet/model/touhou/OperationalData.java
@@ -60,7 +60,7 @@ public class OperationalData extends PagableEntity{
 
     private BigDecimal tradeUserBlance;
 
-    private Integer projectId;
+    private Long projectId;
 
     private String isDelete;
 
@@ -282,11 +282,13 @@ public class OperationalData extends PagableEntity{
 		this.tradeUserBlance = tradeUserBlance;
 	}
 
-	public Integer getProjectId() {
+	
+
+	public Long getProjectId() {
 		return projectId;
 	}
 
-	public void setProjectId(Integer projectId) {
+	public void setProjectId(Long projectId) {
 		this.projectId = projectId;
 	}
